adding pledging and unpledging in mining transaction, Collect Required credits function and few fixes to the previous errors

+++ core/model/mining.go
@@ -15,9 +15,9 @@ type PledgeHistory struct {
 	TransferTokenType  int     `gorm:"column:transfer_tokens_type"`
 	TransferTokenValue float64 `gorm:"column:transfer_token_value"`
 	TransferBlockID    string  `gorm:"column:transfer_block_number_and_id"`
-	Epoch              int     `gorm:"column:epoch"`
-	NextBlockEpoch     int64   `gorm:"column:next_epoch"`
-	TokenCredit        int     `gorm:"column:token_credit"`
+	Epoch              uint64     `gorm:"column:epoch"`
+	NextBlockEpoch     uint64   `gorm:"column:next_epoch"`
+	TokenCredit        uint64  `gorm:"column:token_credit"`
 	TokenCreditStatus  int     `gorm:"column:token_credit_status"`
 }
